caption/requirements.txt
requests
Pillow
numpy
tqdm
argparse
pandas